feat: expose internal components to inject own data
