+++ gw2-ui/src/components/Item/Item.tsx
@@ -1,36 +1,13 @@
-import clsx from 'clsx';
-import { CSSProperties, ReactElement } from 'react';
-import { createItem } from '../../builder';
-import { ItemStatName } from '../../builder/itemStatNames';
+import { ReactElement } from 'react';
 import { useItems } from '../../gw2api/hooks';
-import GW2ApiBackItemDetails from '../../gw2api/types/items/details/backItem';
-import GW2ApiArmorDetails from '../../gw2api/types/items/details/armor';
-import GW2ApiWeaponDetails from '../../gw2api/types/items/details/weapon';
 import GW2ApiItem from '../../gw2api/types/items/item';
-import { capitalize } from '../../helpers/capitalize';
 import Error from '../Error/Error';
 import IconWithText from '../IconWithText/IconWithText';
-import Tooltip, { TooltipProps } from '../Tooltip/Tooltip';
-import WikiLink, { WikiLinkProps } from '../WikiLink/WikiLink';
-import css from './Item.module.css';
-import ItemDetails from './ItemDetails';
+import ItemInternal, { ItemInternalProps } from './ItemInternal';
 
-export type ItemUpgrades = (number | [number, number])[]; // ItemId, or [ItemId, amount] for runes
-export interface ItemProps {
+export interface ItemProps
+  extends Omit<ItemInternalProps, 'dataItem' | 'dataUpgrades'> {
   id: number;
-  text?: string | ((text: string) => string);
-  count?: number;
-  stat?: ItemStatName; // Allow explicitly forcing a certain stat on an item. Many items allow multiple stats and dont provide a "default" one
-  disableIcon?: boolean;
-  disableText?: boolean;
-  disableLink?: boolean;
-  disableTooltip?: boolean;
-  inline?: boolean;
-  tooltipProps?: TooltipProps;
-  wikiLinkProps?: WikiLinkProps;
-  upgrades?: ItemUpgrades;
-  style?: CSSProperties;
-  className?: string;
 }
 const SKILL_ERROR_NAMES = {
   404: 'Item Not Found',
@@ -45,9 +22,6 @@ const SKILL_ERROR_MESSAGES = {
 const Item = (props: ItemProps): ReactElement => {
   const {
     id,
-    text: textProps,
-    count = 1,
-    stat: statRaw,
     disableIcon,
     disableText,
     disableLink,
@@ -59,7 +33,6 @@ const Item = (props: ItemProps): ReactElement => {
     style,
     className,
   } = props;
-  const stat = statRaw ? capitalize(statRaw) : undefined;
 
   if (!id) {
     return (
@@ -111,9 +84,6 @@ const Item = (props: ItemProps): ReactElement => {
     );
   }
   const itemdata = items.data[id];
-  const { name, icon, rarity } = itemdata;
-  // evaluate text prop. if its a function, execute it
-  const text = typeof textProps === 'function' ? textProps(name) : textProps;
 
   // TODO redo the typing for details: the type of the details field depends
   //      on what string is supplied via type (Gw2ApiItemType)
@@ -124,8 +94,7 @@ const Item = (props: ItemProps): ReactElement => {
       let id: number;
       let count: number = 1;
       if (Array.isArray(u)) {
-        id = u[0];
-        count = u[1];
+        [id, count] = u;
       } else {
         id = u;
       }
@@ -133,97 +102,8 @@ const Item = (props: ItemProps): ReactElement => {
     }
   }
 
-  // if there is a stat explicitly requested in the props, we need to augment the existing itemdata to include that stat
-  let stattedItemData = itemdata;
-  if (
-    stat &&
-    (itemdata.type === 'Armor' ||
-      itemdata.type === 'Weapon' ||
-      itemdata.type === 'Trinket' ||
-      itemdata.type === 'Back')
-  ) {
-    // adjust the time due to api inconsistencies: Short Bow <==> Longbow
-    let type = itemdata.type === 'Back' ? 'Back Item' : itemdata.details.type;
-    if (type === 'LongBow') type = 'Longbow';
-    else if (type === 'ShortBow') type = 'Short Bow';
-
-    let createdData;
-    try {
-      createdData = createItem({
-        type,
-        stat,
-        weight:
-          itemdata.type === 'Armor' ? itemdata.details.weight_class : undefined,
-      });
-    } catch (e) {
-      console.error(`Error while loading item with id ${id}!\n ${e}`);
-      return (
-        <Error
-          name="Invalid Item"
-          message={`An error occured ${id}: ${e}`}
-          code={404}
-        />
-      );
-    }
-
-    const details:
-      | GW2ApiArmorDetails
-      | GW2ApiWeaponDetails
-      | GW2ApiBackItemDetails = {
-      ...itemdata.details,
-      infix_upgrade: createdData.details.infix_upgrade,
-    };
-
-    stattedItemData = {
-      ...itemdata,
-      // TypeScript does not understand that this must be of the matching type
-      details: details as any,
-    };
-  }
-
   return (
-    <Tooltip
-      content={<ItemDetails item={stattedItemData} upgrades={upgradedata} />}
-      disabled={disableTooltip}
-      {...tooltipProps}
-      containerProps={{
-        ...tooltipProps?.containerProps,
-        className: clsx(css.container, tooltipProps?.containerProps?.className),
-      }}
-    >
-      <IconWithText
-        icon={icon}
-        text={
-          <>
-            {count > 1 && `${count} `}
-            {disableLink ? (
-              text || name
-            ) : (
-              <WikiLink
-                to={name}
-                text={text}
-                {...wikiLinkProps}
-                className={clsx(
-                  wikiLinkProps?.className,
-                  css[`colorRarity${capitalize(rarity)}`],
-                )}
-              />
-            )}
-          </>
-        }
-        disableIcon={disableIcon}
-        disableText={disableText}
-        inline={inline}
-        iconProps={{
-          applyCount: count,
-          applyCountProps: {
-            className: css.iconApplyCount,
-          },
-        }}
-        style={style}
-        className={clsx(className, css[`colorRarity${capitalize(rarity)}`])}
-      />
-    </Tooltip>
+    <ItemInternal {...props} dataItem={itemdata} dataUpgrades={upgradedata} />
   );
 };
