+++ gw2-ui/src/components/Trait/Trait.tsx
@@ -1,32 +1,12 @@
-import React, { MouseEventHandler, ReactElement } from 'react';
+import { ReactElement } from 'react';
 
-import Tooltip, { TooltipProps } from '../Tooltip/Tooltip';
-import TooltipContainer from '../TooltipContainer/TooltipContainer';
-import IconWithText from '../IconWithText/IconWithText';
-import WikiLink, { WikiLinkProps } from '../WikiLink/WikiLink';
-import AbilityDetails from '../AbilityDetails/AbilityDetails';
-import SPECIALIZATIONS from '../../data/specializations';
-import professioncss from '../Profession/professions.module.css';
-import css from './Trait.module.css';
 import { useTrait } from '../../gw2api/hooks';
 import Error from '../Error/Error';
-import clsx from 'clsx';
-import { CSSProperties } from 'react';
+import IconWithText from '../IconWithText/IconWithText';
+import TraitInternal, { TraitInternalProps } from './TraitInternal';
 
-export interface TraitProps {
+export interface TraitProps extends Omit<TraitInternalProps, 'data'> {
   id: number;
-  text?: string;
-  disableIcon?: boolean;
-  disableText?: boolean;
-  disableLink?: boolean;
-  disableTooltip?: boolean;
-  inline?: boolean;
-  tooltipProps?: TooltipProps;
-  wikiLinkProps?: WikiLinkProps;
-  inactive?: boolean;
-  style?: CSSProperties;
-  className?: string;
-  onClick?: MouseEventHandler<HTMLSpanElement>;
 }
 
 const TRAIT_ERROR_NAMES = {
@@ -40,22 +20,7 @@ const TRAIT_ERROR_MESSAGES = {
 };
 
 const Trait = (props: TraitProps): ReactElement => {
-  const {
-    id,
-    text,
-    disableIcon,
-    disableText,
-    disableLink,
-    disableTooltip,
-    inline,
-    tooltipProps,
-    wikiLinkProps,
-    inactive,
-    style,
-    className,
-    onClick,
-  } = props;
-  const trait = useTrait(id);
+  const trait = useTrait(props.id);
 
   if (trait.loading) {
     return <IconWithText {...props} loading />;
@@ -71,68 +36,7 @@ const Trait = (props: TraitProps): ReactElement => {
     );
   }
 
-  const { name, icon, specialization, skills, slot } = trait.data;
-
-  const [profession] = (specialization &&
-    Object.entries(SPECIALIZATIONS).find(([, specializationIds]) =>
-      specializationIds.includes(specialization),
-    )) || [''];
-
-  return (
-    <Tooltip
-      render={
-        <>
-          <TooltipContainer {...(skills && { className: css.ability })}>
-            <AbilityDetails data={trait.data} />
-          </TooltipContainer>
-
-          {skills &&
-            skills.map((skill, index) => (
-              <TooltipContainer
-                key={`${skill.id}-${index}`}
-                className={css.traitSkill}
-              >
-                <AbilityDetails data={skill} />
-              </TooltipContainer>
-            ))}
-        </>
-      }
-      disabled={disableTooltip}
-      {...tooltipProps}
-    >
-      <IconWithText
-        icon={icon}
-        text={
-          disableLink ? (
-            text || name
-          ) : (
-            <WikiLink
-              to={name}
-              text={text}
-              {...wikiLinkProps}
-              className={
-                profession && professioncss[`coloredProfession${profession}`]
-              }
-            />
-          )
-        }
-        disableIcon={disableIcon}
-        disableText={disableText}
-        inline={inline}
-        iconProps={{
-          zoom: 13,
-          inactive,
-          hexagon: slot === 'Minor',
-        }}
-        style={style}
-        className={clsx(
-          className,
-          profession && professioncss[`coloredProfession${profession}`],
-        )}
-        onClick={onClick}
-      />
-    </Tooltip>
-  );
+  return <TraitInternal {...props} data={trait.data} />;
 };
 
 export default Trait;
